+++ packages/generator/include/combo/common/kaleido_setup.h
@@ -1,0 +1,8 @@
+#ifndef COMBO_COMMON_KALEIDO_SETUP_H
+#define COMBO_COMMON_KALEIDO_SETUP_H
+
+#include <combo.h>
+
+int KaleidoSetup_FrameAdvance(Input* input);
+
+#endif /* COMBO_COMMON_KALEIDO_SETUP_H */