+++ packages/generator/include/combo/mm/play.h
@@ -40,15 +40,27 @@ _Static_assert(sizeof(SramContext) == 0x28, "MM SramContext size is wrong");
 
 typedef struct
 {
-    char    unk_000[0x190];
+    View view;
+    u8* iconItemSegment;
+    u8* iconItem24Segment;
+    u8* iconItemAltSegment;
+    u8* iconItemLangSegment;
+    u8* nameSegment;
+    /* 0x17C */ u8 *iconItemVtxSegment;
+    /* 0x180 */ Vtx *itemPageVtx;
+    /* 0x184 */ Vtx *mapPageVtx;
+    /* 0x188 */ Vtx *questPageVtx;
+    /* 0x18C */ Vtx *maskPageVtx;
     /* 0x190 */ Vtx* vtxBuf;
     char    unk_194[0x58];
     u16     state;
-    char    unk_1ee[0x12];
+    u16     debugEditor;
+    u8      bombersNotebookOpen;
+    Vec3f   eye;
     u16     mainState;
     u16     nextPageMode;
-    u16     screenIndex;
-    u16     switchPageTimer;
+    u16     pageIndex;
+    u16     pageSwitchTimer;
     u16     savePromptState;
     /* 0x20C */ f32 unk20C;
     /* 0x210 */ f32 unk210;
@@ -73,17 +85,30 @@ typedef struct
 }
 PauseContext;
 
-ASSERT_OFFSET(PauseContext, unk_000,            0x000);
+ASSERT_OFFSET(PauseContext, view,               0x000);
 ASSERT_OFFSET(PauseContext, state,              0x1ec);
 ASSERT_OFFSET(PauseContext, mainState,          0x200);
-ASSERT_OFFSET(PauseContext, screenIndex,        0x204);
+ASSERT_OFFSET(PauseContext, pageIndex,          0x204);
 ASSERT_OFFSET(PauseContext, savePromptState,    0x208);
 ASSERT_OFFSET(PauseContext, itemDescriptionOn,  0x2b9);
 _Static_assert(sizeof(PauseContext) == 0x2d0, "MM PauseContext size is wrong");
 
+typedef enum GameOverState
+{
+    /*  0 */ GAMEOVER_INACTIVE,
+    /*  1 */ GAMEOVER_DEATH_START,
+    /*  2 */ GAMEOVER_DEATH_WAIT_GROUND, // wait for player to fall and hit the ground
+    /*  3 */ GAMEOVER_DEATH_FADE_OUT,    // wait before fading out
+    /* 20 */ GAMEOVER_REVIVE_START = 20,
+    /* 21 */ GAMEOVER_REVIVE_RUMBLE,
+    /* 22 */ GAMEOVER_REVIVE_WAIT_GROUND, // wait for player to fall and hit the ground
+    /* 23 */ GAMEOVER_REVIVE_WAIT_FAIRY,  // wait for the fairy to rise all the way up out of player's body
+    /* 24 */ GAMEOVER_REVIVE_FADE_OUT     // fade out the game over lights as player is revived and gets back up
+} GameOverState;
+
 typedef struct
 {
-    int unk;
+    u16 state;
 }
 GameOverContext;
 
@@ -125,20 +150,18 @@ typedef struct ActorContext
     /* 0x1E4 */ TitleCardContext titleCtx;
     /* 0x1F4 */ PlayerImpact playerImpact;
     char        unk_208[0x4c];
-    Actor*      elegyStatues[4];
-    char        unk_264[0x4];
-    u8          unk_268;
+    Actor*      elegyStatues[5];
+    u8          isOverrideInputOn;
     char        unk_269[0x1b];
 }
 ActorContext;
 
-ASSERT_OFFSET(ActorContext, actors,         0x010);
-ASSERT_OFFSET(ActorContext, unk_120,        0x120);
-ASSERT_OFFSET(ActorContext, titleCtx,       0x1e4);
-ASSERT_OFFSET(ActorContext, elegyStatues,   0x254);
-ASSERT_OFFSET(ActorContext, unk_264,        0x264);
-ASSERT_OFFSET(ActorContext, unk_268,        0x268);
-ASSERT_OFFSET(ActorContext, unk_269,        0x269);
+ASSERT_OFFSET(ActorContext, actors,            0x010);
+ASSERT_OFFSET(ActorContext, unk_120,           0x120);
+ASSERT_OFFSET(ActorContext, titleCtx,          0x1e4);
+ASSERT_OFFSET(ActorContext, elegyStatues,      0x254);
+ASSERT_OFFSET(ActorContext, isOverrideInputOn, 0x268);
+ASSERT_OFFSET(ActorContext, unk_269,           0x269);
 
 _Static_assert(sizeof(ActorContext) == 0x284,       "MM ActorContext size is wrong");
 
@@ -494,6 +517,7 @@ ASSERT_OFFSET(PlayState, transitionMode,           0x18b4a);
 #define TRANS_TRIGGER_OFF          0x00
 #define TRANS_TRIGGER_START        0x14
 #define TRANS_TRIGGER_END         -0x14
+
 
 _Static_assert(sizeof(PlayState) == 0x19258, "MM PlayState size is wrong");